refactor(dev-preview): extract output processing into sibling module
- Move the handleData body (URL/readiness detection, compile-marker timers,
  output-error classification) into processDevPreviewOutput and classifyExit
  into classifyDevPreviewExit, both in DevPreviewOutputProcessor.ts
- Deps bag (detector, textDecoder, recordSessionDiagnostic, updateSession,
  emitStateChanged, clearCompiling, pollServerReadiness) keeps the import
  arrow one-directional; classifyDevPreviewExit needs no deps
- DevPreviewSessionService keeps thin handleData/classifyExit wrappers;
  behavior and call sites unchanged
