+++ electron/services/DevPreviewSessionService.ts
@@ -7,6 +7,7 @@ import {
   normalizeViteDevCommand,
 } from "./DevPreviewCommandNormalizer.js";
 import { guardedReinstall, resetCrashLoopGuard } from "./DevPreviewCrashLoopGuard.js";
+import { classifyDevPreviewExit, processDevPreviewOutput } from "./DevPreviewOutputProcessor.js";
 import {
   CACHE_DIRS,
   clearCacheDirs,
@@ -59,7 +60,6 @@ import type {
   DevPreviewDestructivePreviewSizesRequest,
 } from "../../shared/types/ipc/devPreview.js";
 import type { DevServerError } from "../../shared/utils/devServerErrors.js";
-import { detectDevServerError } from "../../shared/utils/devServerErrors.js";
 import { CRASH_SIGNALS, EXPECTED_TERMINATION_SIGNALS } from "./pty/terminalForensics.js";
 import { PERF_MARKS } from "../../shared/perf/marks.js";
 import { markPerformance } from "../utils/performance.js";
@@ -109,9 +109,6 @@ const STALE_START_RECOVERY_MS = 10000;
 const STARTUP_REPLAY_DELAY_MS = 1500;
 const REPLAY_HISTORY_MAX_LINES = 300;
 const PTY_SUBMIT_AFTER_SPAWN_MS = 100;
-const COMPILE_ARM_MS = 1000;
-const COMPILE_CLEAR_MS = 1500;
-
 // Bounds for the `lastOutput` activity hint surfaced in the cross-worktree
 // dashboard. Only the buffer tail is scanned (the last line is all we need) and
 // the result is capped so a single pathological line can't bloat the snapshot.
@@ -1665,147 +1662,15 @@ export class DevPreviewSessionService {
     const session = this.sessions.get(sessionKey);
     if (!session || session.terminalId !== id) return;
 
-    if (session.isRunningInstall) return;
-
-    const dataString = typeof data === "string" ? data : this.textDecoder.decode(data);
-    const result = this.detector.scanOutput(dataString, session.buffer);
-    session.buffer = result.buffer;
-
-    let urlJustStarted = false;
-    if (result.url && result.url !== session.url && result.url !== session.pendingUrl) {
-      markPerformance(PERF_MARKS.DEVPREVIEW_URL_DETECTED, {
-        panelId: session.panelId,
-        projectId: session.projectId,
-        terminalId: id,
-        url: result.url,
-      });
-      this.recordSessionDiagnostic(session, {
-        type: "url-detected",
-        url: capDiagnosticText(result.url),
-      });
-
-      session.readinessAbort?.abort();
-      const abort = new AbortController();
-      session.readinessAbort = abort;
-      session.pendingUrl = result.url;
-      // A new URL (e.g. a port change) starts a fresh poll; allow its own
-      // readiness marker to accelerate it again.
-      session.markerSeen = false;
-
-      this.pollServerReadiness(session, result.url, abort.signal, session.generation);
-      urlJustStarted = true;
-    }
-
-    // A framework readiness line ("ready in N ms", "✓ Ready in", "compiled
-    // successfully") confirms the HTTP server is bound. When a poll is already
-    // mid-cycle (sleeping between HEAD attempts), abort it and re-probe now to
-    // skip the remaining poll interval. The poll itself stays the fallback for
-    // unrecognized frameworks, so no marker means no behavior change.
-    if (result.readyMarker && !session.markerSeen && !urlJustStarted && session.pendingUrl) {
-      session.markerSeen = true;
-      session.readinessAbort?.abort();
-      const abort = new AbortController();
-      session.readinessAbort = abort;
-      this.pollServerReadiness(session, session.pendingUrl, abort.signal, session.generation);
-    }
-
-    // Compile marker detection fires whenever the framework emits a
-    // compile-start line (HMR rebuild, initial compilation burst). Uses a
-    // two-stage timer: arm debounce prevents flicker from rapid marker
-    // bursts, auto-clear removes the signal when markers go quiet.
-    if (
-      result.compileMarker &&
-      session.status !== "stopped" &&
-      session.status !== "error" &&
-      !session.isRunningInstall
-    ) {
-      if (session.compiling) {
-        // Already visible — reset the auto-clear timer so the signal
-        // persists through a burst (HMR chains, multi-file saves).
-        if (session.compilingClearTimer !== null) {
-          clearTimeout(session.compilingClearTimer);
-          session.compilingClearTimer = setTimeout(() => {
-            session.compilingClearTimer = null;
-            this.recordSessionDiagnostic(session, { type: "compile-cleared" });
-            this.clearCompiling(session);
-            this.emitStateChanged(session);
-          }, COMPILE_CLEAR_MS);
-        }
-      } else if (session.compilingTimer === null) {
-        // Start the arm debounce — if no additional markers arrive before
-        // COMPILE_ARM_MS, publish the Compiling label.
-        session.compilingTimer = setTimeout(() => {
-          session.compilingTimer = null;
-          if (
-            session.status === "stopped" ||
-            session.status === "error" ||
-            session.isRunningInstall
-          ) {
-            return;
-          }
-          session.compiling = true;
-          this.recordSessionDiagnostic(session, { type: "compile-started" });
-          this.updateSession(session, { phaseLabel: "Compiling" });
-          session.compilingClearTimer = setTimeout(() => {
-            session.compilingClearTimer = null;
-            this.recordSessionDiagnostic(session, { type: "compile-cleared" });
-            this.clearCompiling(session);
-            this.emitStateChanged(session);
-          }, COMPILE_CLEAR_MS);
-        }, COMPILE_ARM_MS);
-      }
-    }
-
-    // A readyMarker during an active compile phase cancels both timers
-    // and clears the label early — the framework signaled completion.
-    if (result.readyMarker && (session.compiling || session.compilingTimer !== null)) {
-      if (session.compilingTimer !== null) {
-        clearTimeout(session.compilingTimer);
-        session.compilingTimer = null;
-      }
-      if (session.compiling) {
-        this.recordSessionDiagnostic(session, { type: "compile-cleared" });
-      }
-      this.clearCompiling(session);
-      this.emitStateChanged(session);
-    }
-
-    if (!result.error) return;
-    const errorKey = `${result.error.type}:${result.error.message}`;
-    if (errorKey === session.lastErrorKey) return;
-    session.lastErrorKey = errorKey;
-    this.recordSessionDiagnostic(session, {
-      type: "output-error",
-      errorType: result.error.type,
-      message: capDiagnosticText(result.error.message),
-    });
-
-    // Cancel any in-flight readiness poll: a server that still answers HEAD on
-    // a half-started port would otherwise resolve to "running" and clobber the
-    // error/installing status we are about to set.
-    session.readinessAbort?.abort();
-    session.readinessAbort = null;
-    session.pendingUrl = null;
-    session.markerSeen = false;
-    this.clearCompiling(session);
-
-    if (result.error.type === "missing-dependencies") {
-      session.needsInstall = true;
-      this.updateSession(session, {
-        status: "installing",
-        error: result.error,
-        isRestarting: false,
-      });
-      return;
-    }
-
-    this.updateSession(session, {
-      status: "error",
-      error: result.error,
-      url: null,
-      predictedUrl: null,
-      isRestarting: false,
-      phaseLabel: undefined,
+    processDevPreviewOutput(session, id, data, {
+      detector: this.detector,
+      textDecoder: this.textDecoder,
+      recordSessionDiagnostic: (target, input) => this.recordSessionDiagnostic(target, input),
+      updateSession: (target, updates) => this.updateSession(target, updates),
+      emitStateChanged: (target) => this.emitStateChanged(target),
+      clearCompiling: (target) => this.clearCompiling(target),
+      pollServerReadiness: (target, url, signal, generation) =>
+        this.pollServerReadiness(target, url, signal, generation),
     });
   }
 
@@ -1814,44 +1679,7 @@ export class DevPreviewSessionService {
     signal: number | undefined,
     recentOutput: string
   ): DevServerError | null {
-    // Signal-based tier first
-    if (signal !== undefined) {
-      // Check OOM output across all crash signals — Node.js heap OOM sends
-      // SIGABRT (6), not SIGKILL (9). Kernel OOM killer uses SIGKILL.
-      if (
-        CRASH_SIGNALS.has(signal) &&
-        /\b(FATAL ERROR|out of memory|heap out of memory|Cannot allocate memory)\b/i.test(
-          recentOutput
-        )
-      ) {
-        return {
-          type: "oom",
-          message: "Dev server was killed — out of memory.",
-        };
-      }
-      if (signal === 9) {
-        // SIGKILL without OOM output: fall through to output/exit-code tiers
-      } else if (CRASH_SIGNALS.has(signal)) {
-        return {
-          type: "process-crash",
-          message: `Dev server crashed (signal ${signal}).`,
-        };
-      }
-      if (EXPECTED_TERMINATION_SIGNALS.has(signal)) {
-        return null;
-      }
-    }
-
-    // Output-based tier
-    const outputError = detectDevServerError(recentOutput);
-    if (outputError) return outputError;
-
-    // Exit-code fallback
-    if (exitCode === 0) return null;
-    return {
-      type: "unknown",
-      message: `Dev server exited with code ${exitCode}`,
-    };
+    return classifyDevPreviewExit(exitCode, signal, recentOutput);
   }
 
   private handleExit(id: string, exitCode: number, signal?: number): void {
